FEN Parsen beginnt..

+++ src/types.hpp
@@ -77,7 +77,7 @@ struct GameState{
   Color side_to_move;
   CastlingRight castling_rights;
   Square ep_square;
-  uint8_t half_move_clock;
+  uint16_t half_move_clock;
   uint8_t total_move_count;
   Piece captured_piece;
 };
